+++ src/spatialdata_plot/pl/utils.py
@@ -40,7 +40,7 @@ from spatialdata import transform
 from spatialdata._core.query.relational_query import _locate_value, get_values
 from spatialdata._logging import logger as logging
 from spatialdata._types import ArrayLike
-from spatialdata.models import Image2DModel, SpatialElement
+from spatialdata.models import Image2DModel, Labels2DModel, SpatialElement
 from spatialdata.transformations import get_transformation
 
 from spatialdata_plot._logging import logger
@@ -279,19 +279,21 @@ def _get_extent(
             for images_key in sdata.images:
                 for e_id in element_ids:
                     if images_key == e_id:
-                        if not isinstance(sdata.images[e_id], msi.multiscale_spatial_image.MultiscaleSpatialImage):
+                        if isinstance(sdata.images[e_id], spatial_image.SpatialImage):
                             extent[cs_name][e_id] = _get_extent_after_transformations(sdata.images[e_id], cs_name)
                         else:
-                            pass
+                            img = Image2DModel.parse(sdata.images[e_id]["scale0"].ds.to_array().squeeze(axis=0))
+                            extent[cs_name][e_id] = _get_extent_after_transformations(img, cs_name)
 
         if has_labels and cs_contents.query(f"cs == '{cs_name}'")["has_labels"][0]:
             for labels_key in sdata.labels:
                 for e_id in element_ids:
                     if labels_key == e_id:
-                        if not isinstance(sdata.labels[e_id], msi.multiscale_spatial_image.MultiscaleSpatialImage):
+                        if isinstance(sdata.labels[e_id], spatial_image.SpatialImage):
                             extent[cs_name][e_id] = _get_extent_after_transformations(sdata.labels[e_id], cs_name)
                         else:
-                            pass
+                            label = Labels2DModel.parse(sdata.labels[e_id]["scale0"].ds.to_array().squeeze(axis=0))
+                            extent[cs_name][e_id] = _get_extent_after_transformations(label, cs_name)
 
         if has_shapes and cs_contents.query(f"cs == '{cs_name}'")["has_shapes"][0]:
             for shapes_key in sdata.shapes:
@@ -314,11 +316,13 @@ def _get_extent(
 
                         # Split by Point and Polygon:
                         tmp_points = sdata.shapes[e_id][
-                            sdata.shapes[e_id]["geometry"].apply(lambda geom: geom.geom_type == "Point")
+                            sdata.shapes[e_id]["geometry"].apply(
+                                lambda geom: (geom.geom_type == "Point" and not geom.is_empty)
+                            )
                         ]
                         tmp_polygons = sdata.shapes[e_id][
                             sdata.shapes[e_id]["geometry"].apply(
-                                lambda geom: geom.geom_type in ["Polygon", "MultiPolygon"]
+                                lambda geom: (geom.geom_type in ["Polygon", "MultiPolygon"] and not geom.is_empty)
                             )
                         ]
 
@@ -483,23 +487,28 @@ class OutlineParams:
 
     outline: bool
     outline_color: str | list[float]
-    gap_size: float
-    bg_size: float
+    linewidth: float
 
 
 def _set_outline(
     size: float,
     outline: bool = False,
-    outline_width: tuple[float, float] = (0.3, 0.05),
+    outline_width: float = 1.5,
     outline_color: str | list[float] = "#0000000ff",  # black, white
     **kwargs: Any,
 ) -> OutlineParams:
-    bg_width, gap_width = outline_width
-    point = np.sqrt(size)
-    gap_size = (point + (point * gap_width) * 2) ** 2
-    bg_size = (np.sqrt(gap_size) + (point * bg_width) * 2) ** 2
-    # the default black and white colors can be changes using the contour_config parameter
+    # Type checks for outline_width
+    if isinstance(outline_width, int):
+        outline_width = float(outline_width)
+    if not isinstance(outline_width, float):
+        raise TypeError(f"Invalid type of `outline_width`: {type(outline_width)}, expected `float`.")
+    if outline_width == 0.0:
+        outline = False
+    if outline_width < 0.0:
+        logging.warning(f"Negative line widths are not allowed, changing {outline_width} to {(-1)*outline_width}")
+        outline_width = (-1) * outline_width
 
+    # the default black and white colors can be changed using the contour_config parameter
     if (len(outline_color) == 3 or len(outline_color) == 4) and all(isinstance(c, float) for c in outline_color):
         outline_color = matplotlib.colors.to_hex(outline_color)
 
@@ -507,7 +516,7 @@ def _set_outline(
         kwargs.pop("edgecolor", None)  # remove edge from kwargs if present
         kwargs.pop("alpha", None)  # remove alpha from kwargs if present
 
-    return OutlineParams(outline, outline_color, gap_size, bg_size)
+    return OutlineParams(outline, outline_color, outline_width)
 
 
 def _get_subplots(num_images: int, ncols: int = 4, width: int = 4, height: int = 3) -> plt.Figure | plt.Axes: